Update README.md
This tutorial aims to guide the user in creating an EC2 instance with Amazon Linux 2023, configuring an IAM role to allow access to SSM (Systems Manager), and creating a security group to enable ports 8080 and 443. Additionally, the tutorial also configures the EC2 instance to install Docker, Git, Node, and npm, and creates a swap to improve system performance. By the end of the tutorial, the user will have a functioning EC2 instance with a basic security configuration and necessary tools for application development.